--- client/src/components/StateSnapshots.jsx
@@ -1,12 +1,0 @@
-//the name and a simple graphic that will show on the
-//user's profile pages as a simple overview of the saved state
-
-import React from 'react'
-
-export default function StateSnapshot() {
-    return (
-        "hello world"
-    )
-}
-
-//<button onClick={() => (this.props.handleDelete)}>Delete</button>
