+++ client/src/components/user/Register.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import UserForm from './UserForm'
+import UserForm from './UserForm.jsx'
 
 class Register extends React.Component {
     constructor() {
